Amend ForgeVersionHelper to new URL and broken entry
Forge recently moved around some stuff, such as moving the maven to
their own subdomain and the location of the promitions file being slightly changed.

In additions, the promotions file seems to be updated to have uniform entries, removing
the two broken entries from before. However, version `1.17.10_pre4-latest` causes
issue with our semver class, so it's removed now, and maybe the semver class will be
fixed in the future.