Fix staff sign-off code causing out-of-bounds error
The original code queried the first item in the users without checking
if the list has any contents. This led to an IndexOutOfBoundsException
when querying an empty list, and also means that if the staff member used
another emote other than the first one in the list (which _may_ be
possible depending on configuration), it wouldn't register the staff
sign-off.