Add structure for SQL database
To ease the pain of manually writing JDBC code for everything, we use
Jdbi as an intermediate library between JDBC and us. While using a full
ORM (like Hibernate) would be a possible solution, in my opinion, it
would be too heavy for our use.

Flyway's Java API is added to handle migrations of the database in
the actual production environment, as opposed to the Gradle plugin
which is for development instead.

The sticky roles (what I call role persistence) and user join times
data are migrated automatically on launch if the relevant files exist,
then they are renamed to not trigger migration on next boot but still
keep the data just in case there are problems with the migration.