Fix requests not being deleted when intermediate messages are present
For example, if user A posted a request, then user B posted a request,
 then user A left such that their request would be deleted, the request
would not be deleted because the takeWhileAsync predicate would not match
the request from user B (as it's not from the target user, user A).

To fix this, the message filtering is moved to after the messages
within the deletion period have been collected, then filtered for requests
from the target user.