Fix staff sign-off logic in requests deletion code
Because the two conditions of "not having staff sign-off" and "not yet
logged" were combined into one, a message which met the threshold, was
logged before, but _hasn't_ got staff sign-off is deleted because it
doesn't match the combined condition.

Splitting these into two separate conditions, one nested in the other,
fixes this issue. Now, it will _always_ stop early if it detects no
staff sign-off, but it will also log if it hasn't yet been logged about.