Improve text channel use utilities
They now accept the channel key for the configuration directly, and has
callbacks for when a channel is either non-configured or configured but cannot be found.

The most simplified overload with only the key and channel has a default
callback to log the key and channel ID if the channel was configured
but not found.

This should allow better debugging for misconfigured channels for
e.g. update notifications.